ref: stop using try..catch as control flow in Wrapper
Previously, Wrapper used try..catch around a call to useStoreApi to
check whether there was an available StoreContext, instead of simply
using useContext and checking the value. This leads to poor developer
experience when using the devtools feature "pause on caught exceptions"
as it causes the devtools to break on every render of a Wrapper that is
not within a StoreContext provider. The catch is also universal, meaning
that it could catch a completely unrelated error and happily keep going
instead of crashing which is undesirable.

This commit changes this try..catch into a simple call to useContext to
check if a StoreContext is available.