fix: minimize inline style usage
In some environments, inline styles are not permitted due to security
issues (e.g. strict CSP). Some of the current SvelteFlow components set
string values to `style` attribute to apply dynamic styles, which makes
part of the functionalities not available in these environments.

This commit tries to minimize usage of inline styles by

- Use `style:` directive where possible. These are compiled to setting
  properties of the CSSOM, which is permitted by CSP.
- For props that can be set to `style` attribute, default to `undefined`
  instead of empty string to avoid unnecessarily setting the attribute.